+++ src/components/Quote/Quote.tsx
@@ -1,74 +1,16 @@
-import {
-  ExamIc,
-  CommentIc,
-  HeartIc,
-  ShareIc,
-  StarIc,
-} from "../../asset/icon/index";
-
-import React, { useEffect,useState, useRef } from 'react';
 import styled from "styled-components";
 import tw from "twin.macro";
 
 import { Link } from "react-router-dom";
+import ReactItem from "./ReactItem";
+import ProfileImg from "./ProfileImg";
 
 export default function Quote() {
-/*   const componentRef = useRef<HTMLDivElement>(null);
-  const [scrollPosition, setScrollPosition] = useState<number>(0);
-  function handleScroll() {
-    const currentPosition = window.scrollY;
-    setScrollPosition(currentPosition);
-  };
-
-  window.addEventListener('scroll', handleScroll);
-
-  useEffect(() => {
-    const element = componentRef.current;
-    const rect = element?.getBoundingClientRect();
-    console.log( rect?.y);
-    console.log(element);
-   
-  }, [Element, scrollPosition]);
-   */
-
-  const wrapperRef = useRef<HTMLDivElement>(null);
-  const [scrollPosition, setScrollPosition] = useState<number>(0);
-  const [initialOffset, setInitialOffset] = useState<number>(0);
-
-  function handleScroll() {
-    const currentPosition = window.scrollY;
-    setScrollPosition(currentPosition);
-  }
-
-  useEffect(() => {
-    const wrapperElement = wrapperRef.current;
-    const wrapperRect = wrapperElement?.getBoundingClientRect();
-    if (wrapperRect && wrapperElement) {
-      if (initialOffset === 0) {
-        setInitialOffset(wrapperRect.top);
-      }
-      const yOffset = scrollPosition - initialOffset;
-      wrapperElement.style.transform = `translateY(${scrollPosition}px)`;
-     
-    }
-  }, [scrollPosition, initialOffset, Element]);
-
-  useEffect(() => {
-    window.addEventListener('scroll', handleScroll);
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, []);
-
 
   return (
     <QuoteWrapper>
       <QuoteContainer>
-      <div ref={wrapperRef} style={{ position: 'relative', transition: 'transform 0.3s' }}>
-        <UserWrapper>
-          <ExamIc />
-        </UserWrapper>
-        </div>
+     <ProfileImg/>
         <TextWrapper>
           <UserLink to="/user">얏호</UserLink>
           <FollowBtn>팔로우</FollowBtn>
@@ -87,12 +29,7 @@ export default function Quote() {
             <ReactionBtn>
               반응 <ReactNum>123</ReactNum> 개
             </ReactionBtn>
-            <ReactItem>
-              <ReactIcon as={ShareIc} />
-              <ReactIcon as={CommentIc} />
-              <ReactIcon as={StarIc} />
-              <ReactIcon as={HeartIc} />
-            </ReactItem>
+            <ReactItem/>
           </FeedbackWrapper>
         </TextWrapper>
       </QuoteContainer>
@@ -158,14 +95,3 @@ const ReactNum = styled.span`
   ${tw`text-white font-semibold`}
 `;
 
-const ReactItem = styled.span`
-  ${tw`text-gray-400 ml-auto inline-flex items-center leading-none text-sm`}
-`;
-
-const ReactIcon = styled.div`
-  ${tw`
-    w-6
-    h-6
-    ml-3 cursor-pointer`}
-  fill: #d3d3d3;
-`;
